Subiendo ejercico abtract factory y method factory

+++ Factory/Operacion.java
@@ -1,3 +1,3 @@
 public interface Operacion {
-    int operar(int[] numeros);
+    float operar(float[] numeros);
 }
